Refactor adapter models registry module
