+++ tests/adapter_models/__init__.py
@@ -11,38 +11,10 @@ Email: timurkady@yandex.com
 
 from __future__ import annotations
 
-from types import SimpleNamespace
+from .registry import AdapterModelsRegistry, models, registry
 
-from freeadmin.apps.system import (
-    AdminContentType,
-    AdminGroup,
-    AdminGroupPermission,
-    AdminUser,
-    AdminUserPermission,
-    SystemSetting,
-)
-
-
-class AdapterModelsRegistry:
-    """Expose adapter-backed models for tests and fixtures."""
-
-    def __init__(self) -> None:
-        """Bind adapter models into a convenient namespace."""
-
-        self.models = SimpleNamespace(
-            user=AdminUser,
-            user_permission=AdminUserPermission,
-            group=AdminGroup,
-            group_permission=AdminGroupPermission,
-            content_type=AdminContentType,
-            system_setting=SystemSetting,
-        )
-
-
-registry = AdapterModelsRegistry()
-models = registry.models
-
-__all__ = ["registry", "models"]
+__all__ = ["AdapterModelsRegistry", "registry", "models"]
 
 
 # The End
+
